Improve architecture to allow different outputs depending on the configuration of the logger
