+++ src/domain/entities/LoggerConfig.ts
@@ -1,16 +1,27 @@
 import { Id } from "./Base";
 
-type Auth = {
-    username: string;
-    password: string;
-};
+interface LoggerConfigBase {
+    debug?: boolean;
+}
 
-export type LoggerConfig = {
+export type LoggerConfig = ProgramLoggerConfig | ConsoleLoggerConfig;
+
+export interface ConsoleLoggerConfig extends LoggerConfigBase {
+    type: "console";
+}
+
+export interface ProgramLoggerConfig extends LoggerConfigBase {
+    type: "program";
     baseUrl: string;
     auth: Auth;
     programId: Id;
     loggerDataElements: LoggerDataElements;
     organisationUnitId: Id;
+}
+
+type Auth = {
+    username: string;
+    password: string;
 };
 
 export type LoggerDataElements = {
