+++ src/domain/repositories/ProgramRepository.ts
@@ -1,0 +1,6 @@
+import { FutureData } from "../../data/api-futures";
+import { LoggerConfig } from "../../domain/entities/LoggerConfig";
+
+export interface ProgramRepository {
+    checkConfig(config: LoggerConfig): FutureData<boolean>;
+}
