+++ src/domain/usecases/LogMessageUseCase.ts
@@ -2,10 +2,10 @@ import { FutureData } from "../../data/api-futures";
 import { Log } from "../../domain/entities/Log";
 import { LoggerRepository } from "../../domain/repositories/LoggerRepository";
 
-export class LogDebugMessageUseCase {
+export class LogMessageUseCase {
     constructor(private loggerRepository: LoggerRepository) {}
 
-    public execute(log: Log): FutureData<void> {
-        return this.loggerRepository.log(log);
+    public execute(log: Log, isDebug?: boolean): FutureData<void> {
+        return this.loggerRepository.log(log, isDebug);
     }
 }
