--- src/domain/usecases/LogWarnMessageUseCase.ts
@@ -1,11 +1,0 @@
-import { FutureData } from "../../data/api-futures";
-import { Log } from "../../domain/entities/Log";
-import { LoggerRepository } from "../../domain/repositories/LoggerRepository";
-
-export class LogWarnMessageUseCase {
-    constructor(private loggerRepository: LoggerRepository) {}
-
-    public execute(log: Log): FutureData<void> {
-        return this.loggerRepository.log(log);
-    }
-}
